file refactoring and packaging
