+++ simulation_server/lattices/diag0_reconstruction.json
@@ -1,0 +1,67 @@
+{
+    "version": "cheetah-0.7",
+    "title": "unnamed_element_4764",
+    "info": "This is a placeholder lattice description",
+    "root": "unnamed_element_4764",
+    "elements": {
+        "qdg004": ["Quadrupole", {"length": 0.1080000028014183, "k1": 8.77167797088623, "misalignment": [0.0, 0.0], "tilt": 0.0}],
+        "ddg004a": ["Drift", {"length": 0.2705000042915344, "tracking_method": "cheetah"}],
+        "xcdg003": ["HorizontalCorrector", {"length": 0.0, "angle": 0.0}],
+        "ycdg003": ["VerticalCorrector", {"length": 0.0, "angle": 0.0}],
+        "ddg004b": ["Drift", {"length": 0.2705000042915344, "tracking_method": "cheetah"}],
+        "qdg005": ["Quadrupole", {"length": 0.1080000028014183, "k1": -12.053189277648926, "misalignment": [0.0, 0.0], "tilt": 0.0}],
+        "ddg005a": ["Drift", {"length": 0.29600000381469727, "tracking_method": "cheetah"}],
+        "xcdg005": ["HorizontalCorrector", {"length": 0.0, "angle": 0.0}],
+        "ycdg005": ["VerticalCorrector", {"length": 0.0, "angle": 0.0}],
+        "ddg005b": ["Drift", {"length": 0.29600000381469727, "tracking_method": "cheetah"}],
+        "qdg006": ["Quadrupole", {"length": 0.1080000028014183, "k1": 6.830639839172363, "misalignment": [0.0, 0.0], "tilt": 0.0}],
+        "ddg006a": ["Drift", {"length": 0.7327514290809631, "tracking_method": "cheetah"}],
+        "tcydg0": ["Drift", {"length": 1.600000023841858, "tracking_method": "cheetah"}],
+        "ddg006b": ["Drift", {"length": 0.12714000046253204, "tracking_method": "cheetah"}],
+        "bpmdg0rf": ["Marker", {}],
+        "ddg006c": ["Drift", {"length": 0.17285999655723572, "tracking_method": "cheetah"}],
+        "tcxdg0": ["Drift", {"length": 1.600000023841858, "tracking_method": "cheetah"}],
+        "ddg006d": ["Drift", {"length": -1.696751356124878, "tracking_method": "cheetah"}],
+        "qdg007": ["Quadrupole", {"length": 0.1080000028014183, "k1": -4.543712615966797, "misalignment": [0.0, 0.0], "tilt": 0.0}],
+        "ddg007": ["Drift", {"length": 0.5479999780654907, "tracking_method": "cheetah"}],
+        "qdg008": ["Quadrupole", {"length": 0.1080000028014183, "k1": 11.37702751159668, "misalignment": [0.0, 0.0], "tilt": 0.0}],
+        "ddg008a": ["Drift", {"length": 0.29600000381469727, "tracking_method": "cheetah"}],
+        "xcdg008": ["HorizontalCorrector", {"length": 0.0, "angle": 0.0}],
+        "ycdg008": ["VerticalCorrector", {"length": 0.0, "angle": 0.0}],
+        "ddg008b": ["Drift", {"length": 0.29600000381469727, "tracking_method": "cheetah"}],
+        "qdg009": ["Quadrupole", {"length": 0.1080000028014183, "k1": -11.539229393005371, "misalignment": [0.0, 0.0], "tilt": 0.0}],
+        "ddg009a": ["Drift", {"length": 0.29600000381469727, "tracking_method": "cheetah"}],
+        "otrdg01": ["Drift", {"length": 0.0, "tracking_method": "cheetah"}],
+        "otrdg01_yag": ["Drift", {"length": 0.0, "tracking_method": "cheetah"}],
+        "ddg009b": ["Drift", {"length": 0.699999988079071, "tracking_method": "cheetah"}],
+        "rfbdg002": ["Drift", {"length": 0.0, "tracking_method": "cheetah"}],
+        "ddg009c": ["Drift", {"length": 0.30000001192092896, "tracking_method": "cheetah"}],
+        "otrdg02": ["Screen", {"resolution": [1024, 1024], "pixel_size": [1.1000000085914508e-05, 1.1000000085914508e-05], "binning": 1, "misalignment": [0.0, 0.0], "method": "kde", "kde_bandwidth": 1.1000000085914508e-05, "is_active": true}],
+        "ddg009d": ["Drift", {"length": 0.30000001192092896, "tracking_method": "cheetah"}],
+        "wsdg01": ["Marker", {}],
+        "ddg009e": ["Drift", {"length": 0.699999988079071, "tracking_method": "cheetah"}],
+        "otrdg03": ["Drift", {"length": 0.0, "tracking_method": "cheetah"}],
+        "otrdg03_yag": ["Drift", {"length": 0.0, "tracking_method": "cheetah"}],
+        "ddg009f": ["Drift", {"length": 0.6660000085830688, "tracking_method": "cheetah"}],
+        "qdg010": ["Quadrupole", {"length": 0.1080000028014183, "k1": 13.016651153564453, "misalignment": [0.0, 0.0], "tilt": 0.0}],
+        "ddg010a": ["Drift", {"length": 0.29600000381469727, "tracking_method": "cheetah"}],
+        "xcdg010": ["HorizontalCorrector", {"length": 0.0, "angle": 0.0}],
+        "ycdg010": ["VerticalCorrector", {"length": 0.0, "angle": 0.0}],
+        "ddg010b": ["Drift", {"length": 0.29600000381469727, "tracking_method": "cheetah"}],
+        "qdg011": ["Quadrupole", {"length": 0.1080000028014183, "k1": -10.983651161193848, "misalignment": [0.0, 0.0], "tilt": 0.0}],
+        "ddg011a": ["Drift", {"length": 0.30000001192092896, "tracking_method": "cheetah"}],
+        "rfbdg003": ["Drift", {"length": 0.0, "tracking_method": "cheetah"}],
+        "ddg011b": ["Drift", {"length": 1.1047999858856201, "tracking_method": "cheetah"}],
+        "bydg0": ["Dipole", {"length": 0.5435000061988831, "angle": -0.6108652353286743, "k1": 0.0, "dipole_e1": -0.12723450362682343, "dipole_e2": -0.12723450362682343, "tilt": 1.5707963705062866, "gap": 0.03400000184774399, "gap_exit": 0.03400000184774399, "fringe_integral": 0.39100000262260437, "fringe_integral_exit": 0.39100000262260437, "fringe_at": "both", "fringe_type": "linear_edge", "tracking_method": "cheetah"}],
+        "ddg012a": ["Drift", {"length": 0.3952963054180145, "tracking_method": "cheetah"}],
+        "bpmdg012": ["Marker", {}],
+        "ddg012b": ["Drift", {"length": 0.23999999463558197, "tracking_method": "cheetah"}],
+        "otrdg04": ["Screen", {"resolution": [1024, 1024], "pixel_size": [1.1000000085914508e-05, 1.1000000085914508e-05], "binning": 1, "misalignment": [0.0, 0.0], "method": "kde", "kde_bandwidth": 1.1000000085914508e-05, "is_active": true}],
+        "ddg012c": ["Drift", {"length": 0.30000001192092896, "tracking_method": "cheetah"}],
+        "fcdg0du": ["Marker", {}],
+        "enddiag0": ["Marker", {}]
+    },
+    "lattices": {
+        "unnamed_element_4764": ["qdg004", "ddg004a", "xcdg003", "ycdg003", "ddg004b", "qdg005", "ddg005a", "xcdg005", "ycdg005", "ddg005b", "qdg006", "ddg006a", "tcydg0", "ddg006b", "bpmdg0rf", "ddg006c", "tcxdg0", "ddg006d", "qdg007", "ddg007", "qdg008", "ddg008a", "xcdg008", "ycdg008", "ddg008b", "qdg009", "ddg009a", "otrdg01", "otrdg01_yag", "ddg009b", "rfbdg002", "ddg009c", "otrdg02", "ddg009d", "wsdg01", "ddg009e", "otrdg03", "otrdg03_yag", "ddg009f", "qdg010", "ddg010a", "xcdg010", "ycdg010", "ddg010b", "qdg011", "ddg011a", "rfbdg003", "ddg011b", "bydg0", "ddg012a", "bpmdg012", "ddg012b", "otrdg04", "ddg012c", "fcdg0du", "enddiag0"]
+    }
+}
