+++ simulation_server/virtual_accelerator/__init__.py
@@ -1,0 +1,3 @@
+from simulation_server.virtual_accelerator.virtual_accelerator import VirtualAccelerator
+
+__all__ = ["VirtualAccelerator"]
